requirements.txt
Flask==0.11.1
requests==2.9.1
boto==2.41.0
pyzmq==15.3.0